Attempt t clean up
Wait, are you telling me that I just missed some ' mark in da73206 and
because of it for days thought that the most obvious solution doesn't
work?

Cleaning up. There is some suspicion that -A option in cmake file won't
work on linux, but point is to ignore it there.
Lets see.